pyproject.toml
[project]
name = "events-aggregator"
version = "0.1.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "cachetools>=5.5.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.6.0", "pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]